+++ reconstruction/dc/pom.xml
@@ -37,9 +37,15 @@
       <version>4.3-SNAPSHOT</version>
     </dependency>
 
-	<dependency>
+    <dependency>
       <groupId>org.jlab.clas</groupId>
       <artifactId>clas-reco</artifactId>
+      <version>5.0-SNAPSHOT</version>
+    </dependency>
+
+    <dependency>
+      <groupId>org.jlab.clas</groupId>
+      <artifactId>clas-analysis</artifactId>
       <version>5.0-SNAPSHOT</version>
     </dependency>
